Changes after Te-Wei's comments.
