+++ src/tcpip.lvproj
@@ -169,9 +169,8 @@
 				<Item Name="readKeysFromConfigFile.vi" Type="VI" URL="../../simulator/configFile/readKeysFromConfigFile.vi"/>
 				<Item Name="readConfigEvtTel.vi" Type="VI" URL="../../simulator/configFile/readConfigEvtTel.vi"/>
 			</Item>
-			<Item Name="ComponentSimulatorsLauncher.vi" Type="VI" URL="../../simulator/ComponentSimulatorsLauncher.vi"/>
-			<Item Name="ComponentSimulatorA.vi" Type="VI" URL="../../simulator/ComponentSimulatorA.vi"/>
-			<Item Name="ComponentSimulatorB.vi" Type="VI" URL="../../simulator/ComponentSimulatorB.vi"/>
+			<Item Name="ComponentSimulatorLauncher.vi" Type="VI" URL="../../simulator/ComponentSimulatorLauncher.vi"/>
+			<Item Name="ComponentSimulator.vi" Type="VI" URL="../../simulator/ComponentSimulator.vi"/>
 		</Item>
 		<Item Name="FactoryTcpServer" Type="Folder">
 			<Item Name="FactoryTcpServer.lvclass" Type="LVClass" URL="../FactoryTcpServer/FactoryTcpServer.lvclass"/>
@@ -635,7 +634,7 @@
 				<Property Name="Bld_previewCacheID" Type="Str">{5F67191D-C3BD-11EB-91FF-02428DC4B39C}</Property>
 				<Property Name="Bld_userLogFile" Type="Path">../build/buildComponentSimulator.txt</Property>
 				<Property Name="Bld_userLogFile.pathType" Type="Str">relativeToCommon</Property>
-				<Property Name="Bld_version.build" Type="Int">32</Property>
+				<Property Name="Bld_version.build" Type="Int">34</Property>
 				<Property Name="Bld_version.major" Type="Int">1</Property>
 				<Property Name="DestinationCount" Type="Int">2</Property>
 				<Property Name="Destination[0].destName" Type="Str">ComponentSimulator</Property>
@@ -644,15 +643,18 @@
 				<Property Name="Destination[0].type" Type="Str">App</Property>
 				<Property Name="Destination[1].destName" Type="Str">Support Directory</Property>
 				<Property Name="Destination[1].path" Type="Path">../build/data</Property>
-				<Property Name="SourceCount" Type="Int">3</Property>
-				<Property Name="Source[0].itemID" Type="Str">{A4DE72DD-2F81-11ED-A973-B04F130BB390}</Property>
+				<Property Name="SourceCount" Type="Int">4</Property>
+				<Property Name="Source[0].itemID" Type="Str">{F469E460-D59F-11EE-A973-B04F130BB390}</Property>
 				<Property Name="Source[0].type" Type="Str">Container</Property>
 				<Property Name="Source[1].destinationIndex" Type="Int">0</Property>
-				<Property Name="Source[1].itemID" Type="Ref">/My Computer/simulator/ComponentSimulatorA.vi</Property>
-				<Property Name="Source[1].sourceInclusion" Type="Str">TopLevel</Property>
+				<Property Name="Source[1].itemID" Type="Ref">/My Computer/simulator/ComponentSimulator.vi</Property>
 				<Property Name="Source[1].type" Type="Str">VI</Property>
 				<Property Name="Source[2].destinationIndex" Type="Int">0</Property>
 				<Property Name="Source[2].itemID" Type="Ref">/My Computer/config/default.ini</Property>
+				<Property Name="Source[3].destinationIndex" Type="Int">0</Property>
+				<Property Name="Source[3].itemID" Type="Ref">/My Computer/simulator/ComponentSimulatorLauncher.vi</Property>
+				<Property Name="Source[3].sourceInclusion" Type="Str">TopLevel</Property>
+				<Property Name="Source[3].type" Type="Str">VI</Property>
 				<Property Name="TgtF_fileDescription" Type="Str">ComponentSimulator</Property>
 				<Property Name="TgtF_internalName" Type="Str">ComponentSimulator</Property>
 				<Property Name="TgtF_legalCopyright" Type="Str">Copyright 2021 </Property>
